pyproject.toml
[project]
name = "springx-paradigma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "seaborn>=0.13.2",
    "streamlit>=1.28.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.12.0",
]
